+++ includes/shortcodes/event-search/class-shortcode-admin.php
@@ -41,6 +41,27 @@
 		 * @return array
 		 */
 		function fields( $fields, $button_data ) {
+
+            $fields[] = array(
+                'name' => 'Disable Header',
+                'desc' => 'To disable header',
+                'id'   => 'disable_header',
+                'type' => 'checkbox',
+            );
+
+            $fields[] = array(
+                'name' => 'Disable Nav Bar',
+                'desc' => 'To disable Nav Bar',
+                'id'   => 'disable_nav',
+                'type' => 'checkbox',
+            );
+
+            $fields[] = array(
+                'name' => 'Disable Mini Category List',
+                'desc' => 'To disable mini category list',
+                'id'   => 'disable_cateogy_list',
+                'type' => 'checkbox',
+            );
 			
 			return $fields;
 		}