+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased] - 0.6.6
+## [Unreleased] - 0.6.8
 ### Added
 - Event categories shortcode btn fields
 - Event categories single shortcode btn fields
 - Event partner list shortcode btn fields
+- Event search shortcode btn fields
 
 ### Changed
 - Project menu item in nav menu
@@ -13,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Conditions for shortcode atts events_category
 - Conditions for shortcode atts events_category_single
 - Conditions for shortcode atts events_partner_list
+- Conditions for shortcode atts events_search
 
 ### Fixed
 - Single event page time display format fix